requirements.txt
click
requests
pprint
prettytable
pyyaml
kubernetes
